tools/read_from_dbox.py
import io
import dropbox
from moviepy.editor import VideoFileClip

# Set your Dropbox access token
access_token = "f2vo0ff567k9m5v"

# Initialize the Dropbox client
dbx = dropbox.Dropbox(access_token)

# Specify the Dropbox file path
dropbox_path = "/path/to/your/video.mp4"

# Download the file from Dropbox
_, response = dbx.files_download(dropbox_path)

# Load the video data into a MoviePy VideoFileClip
video_data = io.BytesIO(response.content)
video_clip = VideoFileClip(video_data)

video_clip.write_videofile("output.mp4")

# Use the video_clip in your MoviePy project

