requirements.txt
moviepy
Wand
streamlit
qrcode
Pillow
numpy
pyzbar